ajout git_ignore

+++ Tickets/models.py
@@ -8,7 +8,7 @@ from Etudiants.models import Etudiant
 
 
 class Ticket_Repas(models.Model):
-    nbre_tickets_repas = models.FloatField(default=0.0, null=True, blank=True)
+    nbre_tickets_repas = models.PositiveIntegerField(default=0, null=True, blank=True)
     etudiant = models.ForeignKey(Etudiant,related_name="tickets_repas",on_delete=models.CASCADE)
     
     def save(self, *args, **kwargs):
@@ -26,7 +26,7 @@ class Ticket_Repas(models.Model):
 
 
 class Ticket_Dej(models.Model):
-    nbre_tickets_dej = models.FloatField(default=0.0, null=True, blank=True)
+    nbre_tickets_dej = models.PositiveIntegerField(default=0, null=True, blank=True)
     etudiant = models.ForeignKey(Etudiant,related_name="tickets_dej",on_delete=models.CASCADE)
     
     def save(self, *args, **kwargs):
